[legacy] move outdated zero to legacy (#4696)
