+++ colossalai/legacy/zero/gemini/ophooks/utils.py
