+++ colossalai/nn/layer/__init__.py
@@ -1,2 +1,2 @@
-from .moe import *
+# from .moe import *
 from .utils import *
